+++ client/components/subcomponents/Player.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react'
-import {Text} from 'react-konva'
+import {Text, Circle} from 'react-konva'
 
 export default class Player extends Component {
   constructor(props) {
@@ -12,10 +12,11 @@ export default class Player extends Component {
 
   render() {
     return (
-      <Text
-        text={this.props.text}
+      <Circle
+        // text={this.props.text}
         x={this.state.player.x}
         y={this.state.player.y}
+        radius={7}
         draggable
         fill={this.state.isDragging ? 'black' : this.props.color}
         onDragStart={() => {
