--- client/utils/resizeImage.js
@@ -1,3 +1,0 @@
-export const resizeImage = image => {
-  return image
-}
